chore: move label logic from FieldDropdown to Field
